requirements.txt
colorama==0.3.7
Django==1.9.7
django-crispy-forms==1.6.0
django-filter==0.13.0
django-material==0.8.0
djangorestframework==3.4.0
facebook-sdk==2.0.0
flake8==2.6.2
gunicorn==19.6.0
httpie==0.9.4
jedi==0.9.0
mccabe==0.5.0
pep8==1.7.0
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
setproctitle==1.1.10
six==1.10.0
